requirements.txt
numpy==1.18.5
pandas==1.0.5
matplotlib==3.2.2
seaborn==0.10.1
scikit-learn==0.23.1
streamlit==0.71.0
Cython
scikit-surprise
